cpp_module_04: clean and ready to push.

+++ cpp_module_01/ex01/ZombieHorde.cpp
@@ -3,7 +3,7 @@
 Zombie  *zombieHorde(int N, std::string name) {
     Zombie  *zombie_horde = new Zombie[N];
 
-    for (size_t i = 0; i < N; i++) {
+    for (int i = 0; i < N; i++) {
         zombie_horde[i].setName(name + "_" + std::to_string(i));
     }
     return (zombie_horde);
